test(telemetry): colocate classification tests and trim metadata
